Switch Vagrant provisioner to ansible
Use Vagrant's ansible provisioner directly. This removes the need for a
shell script. Note that this changes where ansible runs to the host
which is more like how it would run in a live environment.

This change highlighted an issue around becoming an unprivileged user,
which is solved by using SSH pipelining.
See: http://docs.ansible.com/ansible/become.html#becoming-an-unprivileged-user